created README
# IntFicPy
I've long had an interest in parser-based interactive fiction, both as a creator, and a consumer. When I started creating games, I found some very good tools, but all of them required learning a completely new programming language. This is my attempt to remove barriers to game creation by simplifying the process of creating interactive fiction with Python. IntFicPy is not ready for real use at this point, but it is possible to use it to build a basic game.

# What's Included?
The current classes included are Player, Room, Verb, and Thing (with subclass Actor). In the near future, the additional Thing subclasses of Surface, Container, Decoration, and Openable will be added.
All standard verbs will be built in. Currently, the verbs completed are go [direction], get/take, drop, look, examine, and ask. More will be added in the near future.

# Installtion and Use
Currently, IntFicPy is incomplete, and likely buggy.