recording - close record file between turns
Instead of keeping the record file open for the entire course of the recording,
open it to write each turn, and close it again.

This prevents data loss in the event of sudden program close or crash.

Handling this properly within the IFP engine removes the need for the GUI
to handle closing the file on program exit.