!refactor: move Thing class-conditional logic from PourIntoVerb to Thing classes
API CHANGE: Liquid.mixWith() method now has a new kwarg "event", that
determines which IFPEvent the text it produces should be added to.
If we adopt this new paradigm, this may  become the standard for
all Thing & subclass methods that print text.

We want to stop verbs checking the Thing class of the items they are called on,
and let each class handle its own logic instead. This is a prototype of how this
might be done.